requirements.txt
asgiref==3.11.0
asttokens==2.4.1
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.2
debugpy==1.8.8
decorator==5.1.1
Django==5.2.9
executing==2.1.0
gunicorn==23.0.0
idna==3.11
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pywin32==308
pyzmq==26.2.0
requests==2.32.5
six==1.16.0
sqlparse==0.5.5
stack-data==0.6.3
stripe==14.1.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.3
urllib3==2.6.2
wcwidth==0.2.13
whitenoise==6.11.0
